requirements.txt
gradio
ultralytics
numpy
pandas
pillow
torch
torchvision
